Horse Race Simulator - V8 May 30, 2023
For some unknown reason, the code will sometimes set horse count to 3 on the second loop.  Sometimes, it won't do this and will instead refuse to heal horses upon starting the game.  I've checked many times, the only place where horses are healed or created is in the start, which it always runs through fully, it just sometimes won't work and sometimes will.  I have absolutely no clue why this glitch is so inconsistent, but it is breaking a lot of code which is really annoying.  I spent an hour with the code adding 3 horses when it shouldn't, then in the final 10 or so minutes of class--without me changing anything--the code started properly adding horses but wouldn't heal them anymore.  This is so stupid.