Horse Race Simulator - V15 June 8, 2023
Added in the ability to steal horses.  Now, when the player is broke they can steal a horse.  This randomly chooses one of the three (or two) available horses and removes them from the game, giving the player a sum of money.  The game will then display the horse as being inactive everywhere, and will move them out of the racetrack.
A lot of this code could be cleaned up, but overall it is functional with no known glitches.