Horse Race Simulator - V17 June 15, 2023
Reworked all of the XML code.  It now works.  It can load a saved game or create a new one.  If a game is loaded but the XML does not exist, it will create a new game to avoid crashing.  Gameplay will also work.  Horses that are injured or inactive will be removed properly.  Injured horses will only be injured for the one match that they are meant to be injured for.
All that is left to do is to add the ability to save the game whenever exiting from the menu screen.