Horse Race Simulator - V19 June 20, 2023
Fixed any problems with saving the game.  Cleaned up/reworded code throughout everything.  Added a cut scene for when there is just one horse in the race.
Stuff that is left to do is sounds, music, cut scene graphics, the shop, and game endings.  I probably won't get it all done in time, so the main goal is to finish the visuals and sounds for everything that already works.